+++ src/sources/sources.module.ts
@@ -2,12 +2,14 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { SourcesController } from './controllers/sources.controller';
 import { Source } from './entities/source.entity';
+import { UserSourcePreference } from './entities/user-source-preference.entity';
 import { SourcesService } from './services/sources.service';
+import { UserSourcePreferenceService } from './services/user-source-preference.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Source])],
+  imports: [TypeOrmModule.forFeature([Source, UserSourcePreference])],
   controllers: [SourcesController],
-  providers: [SourcesService],
-  exports: [SourcesService],
+  providers: [SourcesService, UserSourcePreferenceService],
+  exports: [SourcesService, UserSourcePreferenceService],
 })
 export class SourcesModule {}